Fix feature materialization error due to labeling np.float32 value as PrimitiveFeastType.FLOAT64

+++ featurebyte/feast/enum.py
@@ -29,7 +29,7 @@ def to_feast_primitive_type(dtype: DBVarType) -> PrimitiveFeastType:
     mapping = {
         DBVarType.BOOL: PrimitiveFeastType.BOOL,
         DBVarType.VARCHAR: PrimitiveFeastType.STRING,
-        DBVarType.FLOAT: PrimitiveFeastType.FLOAT64,
+        DBVarType.FLOAT: PrimitiveFeastType.FLOAT32,
         DBVarType.INT: PrimitiveFeastType.INT64,
         DBVarType.TIMESTAMP_TZ: PrimitiveFeastType.UNIX_TIMESTAMP,
         DBVarType.TIMESTAMP: PrimitiveFeastType.UNIX_TIMESTAMP,
